rename client trait and move it from network to client
