+++ node/client/Cargo.toml
@@ -8,13 +8,11 @@ parking_lot = "0.6"
 log = "0.4"
 
 beacon = { path = "../../core/beacon" }
-network = { path = "../network" }
 node-runtime = { path = "../runtime" }
 primitives = { path = "../../core/primitives" }
 storage = { path = "../../core/storage" }
 
 [dev-dependencies]
-network = { path = "../network", features = ["test-utils"]}
 storage = { path = "../../core/storage", features = ["test-utils"] }
 
 [features]
