+++ node/client/src/test_utils.rs
@@ -7,14 +7,18 @@ use Client;
 
 pub fn generate_test_client() -> Client {
     let storage = Arc::new(create_memory_db());
-    let genesis_wasm = include_bytes!(
-        "../../../core/wasm/runtest/res/wasm_with_mem.wasm"
-    ).to_vec();
-    let chain_spec = ChainSpec {
-        balances: vec![],
-        initial_authorities: vec![],
-        genesis_wasm,
-    };
+    let genesis_wasm = include_bytes!("../../../core/wasm/runtest/res/wasm_with_mem.wasm").to_vec();
+    let chain_spec = ChainSpec { balances: vec![], initial_authorities: vec![], genesis_wasm };
     let signer = Arc::new(InMemorySigner::default());
     Client::new(&chain_spec, storage, signer)
+}
+
+impl Client {
+    pub fn num_transactions(&self) -> usize {
+        self.tx_pool.read().len()
+    }
+
+    pub fn num_blocks_in_queue(&self) -> usize {
+        self.import_queue.read().len()
+    }
 }
