+++ node/network/Cargo.toml
@@ -19,6 +19,7 @@ rand = "0.5"
 libp2p = { git = "https://github.com/libp2p/rust-libp2p", rev = "d961e656a74d1bab5366d371a06f9e10d5f4a6c5", default-features = false, features = ["secio-rsa", "secio-secp256k1"] }
 substrate-network-libp2p = { git = "https://github.com/nearprotocol/substrate", rev = "5f1ec2832a9a79506985c003b5c041c760e39c4f" }
 primitives = { path = "../../core/primitives" }
+client = { path = "../client", features = ["test-utils"] }
 
 [features]
 test-utils = []
