+++ node/network/src/protocol.rs
@@ -1,4 +1,4 @@
-use client::Client;
+use client::chain::Chain;
 use io::{NetSyncIo, SyncIo};
 use message::{self, Message, MessageBody};
 use parking_lot::RwLock;
@@ -7,10 +7,10 @@ use primitives::traits::{Block, Decode, Encode, GenericResult, Header as BlockHe
 use primitives::types::{BlockId, SignedTransaction};
 use rand::{seq, thread_rng};
 use std::collections::HashMap;
+use std::sync::Arc;
 use std::time;
 use substrate_network_libp2p::{NodeIndex, ProtocolId, Secret, Severity};
 use test_utils;
-use std::sync::Arc;
 
 /// time to wait (secs) for a request
 const REQUEST_WAIT: u64 = 60;
@@ -73,7 +73,7 @@ pub struct Protocol<B: Block, H: ProtocolHandler> {
     // info about peers
     peer_info: RwLock<HashMap<NodeIndex, PeerInfo>>,
     // backend client
-    client: Arc<Client<B>>,
+    client: Arc<Chain<B>>,
     // callbacks
     handler: Option<Box<H>>,
     // phantom data for keep T
@@ -84,7 +84,7 @@ pub trait ProtocolHandler: 'static {
 }
 
 impl<B: Block, H: ProtocolHandler> Protocol<B, H> {
-    pub fn new(config: ProtocolConfig, handler: H, client: Arc<Client<B>>) -> Protocol<B, H> {
+    pub fn new(config: ProtocolConfig, handler: H, client: Arc<Chain<B>>) -> Protocol<B, H> {
         Protocol {
             config,
             handshaking_peers: RwLock::new(HashMap::new()),
@@ -309,12 +309,12 @@ impl<B: Block, H: ProtocolHandler> Protocol<B, H> {
             .iter()
             .filter_map(|(id, info)| info.request_timestamp.as_ref().map(|x| (id, x)))
             .chain(handshaking_peers.iter())
-            {
-                if (cur_time - *time_stamp).as_secs() > REQUEST_WAIT {
-                    trace!(target: "sync", "Timeout {}", *peer);
-                    aborting.push(*peer);
-                }
+        {
+            if (cur_time - *time_stamp).as_secs() > REQUEST_WAIT {
+                trace!(target: "sync", "Timeout {}", *peer);
+                aborting.push(*peer);
             }
+        }
         for peer in aborting {
             net_sync.report_peer(peer, Severity::Timeout);
         }
@@ -342,7 +342,11 @@ impl<B: Block, H: ProtocolHandler> Protocol<B, H> {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use client::test_utils::*;
+    use primitives::hash::hash;
     use primitives::types;
+    use std::cell::RefCell;
+    use std::rc::Rc;
     use test_utils::*;
 
     impl<B: Block, H: ProtocolHandler> Protocol<B, H> {
@@ -373,5 +377,25 @@ mod tests {
         let protocol = Protocol::new(config, MockProtocolHandler::default(), mock_client);
         let tx = types::SignedTransaction::empty();
         protocol.on_transaction_message(tx);
+    }
+
+    #[test]
+    fn test_protocol_and_client() {
+        let client = Arc::new(generate_test_client());
+        let handler = MockHandler { client: client.clone() };
+        let config = ProtocolConfig::new_with_default_id(special_secret());
+        let protocol = Protocol::new(config, handler, client.clone());
+        let network_service = Rc::new(RefCell::new(default_network_service()));
+        let mut net_sync = NetSyncIo::new(network_service, protocol.config.protocol_id);
+        protocol.on_transaction_message(SignedTransaction::new(
+            123,
+            types::TransactionBody::new(1, hash(b"bob"), hash(b"alice"), 10, String::new(), vec![]),
+        ));
+        assert_eq!(client.num_transactions(), 1);
+        assert_eq!(client.num_blocks_in_queue(), 0);
+        protocol.prod_block::<MockBlockHeader>(&mut net_sync);
+        assert_eq!(client.num_transactions(), 0);
+        assert_eq!(client.num_blocks_in_queue(), 0);
+        assert_eq!(client.best_index(), 1);
     }
 }
